+++ src/main/java/me/nikl/cookieclicker/CookieClicker.java
@@ -1,6 +1,10 @@
 package me.nikl.cookieclicker;
 
+import me.nikl.cookieclicker.data.Database;
+import me.nikl.cookieclicker.data.FileDatabase;
+import me.nikl.cookieclicker.data.MySQLDatabase;
 import me.nikl.gamebox.GameBox;
+import me.nikl.gamebox.GameBoxSettings;
 import me.nikl.gamebox.game.Game;
 import me.nikl.gamebox.game.GameSettings;
 
@@ -10,6 +14,7 @@ import me.nikl.gamebox.game.GameSettings;
  * Main class of the GameBox game Cookie Clicker
  */
 public class CookieClicker extends Game {
+    private Database database;
 
     public CookieClicker(GameBox gameBox) {
         super(gameBox, GameBox.MODULE_COOKIECLICKER);
@@ -18,11 +23,16 @@ public class CookieClicker extends Game {
     @Override
     public void onDisable() {
         ((CCGameManager) gameManager).onShutDown();
+        database.onShutDown();
     }
 
     @Override
     public void init() {
-
+        if (GameBoxSettings.useMysql) {
+            this.database = new MySQLDatabase(this);
+        } else {
+            this.database = new FileDatabase(this);
+        }
     }
 
     @Override
@@ -40,5 +50,9 @@ public class CookieClicker extends Game {
     @Override
     public void loadGameManager() {
         this.gameManager = new CCGameManager(this);
+    }
+
+    public Database getDatabase() {
+        return this.database;
     }
 }